+++ src/stores/orders/orderSlice.ts
@@ -195,9 +195,10 @@ const orderSlice = createSlice({
     hideSnackbar: (state) => ({
       ...state,
       snackbar: {
+        ...state.snackbar,
         open: false,
-        message: '',
-        severity: 'info',
+        // message: '',
+        // severity: 'info',
       },
     }),
   },
